+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fix security bundle salt configuration for itsdangerous 2.0+
 - add shell readline completion (from [Flask PR 3960](https://github.com/pallets/flask/pull/3960))
 - fix `BundleBlueprint.register` to work with Flask 2.0+
+- fix compatibility with click 8.0+
 - bump required flask-sqlalchemy-unchained version
 - API bundle fixes and improvements